tong hop

+++ src/ra/business/IEmployee.java
@@ -3,7 +3,7 @@ package ra.business;
 import java.util.Scanner;
 
 public interface IEmployee {
-    public static final long BASIC_SALARY=1300000;
+    public static final double BASIC_SALARY=1300000;
     public void inputData(Scanner sc);
     public void displayData();
 }
